mobile-guide/backend/gait-analysis/requirements.txt
# Flask and web framework
Flask==3.1.0
Flask-CORS==5.0.0

# Scientific computing
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.0.0

# HTTP requests for dataset download
requests>=2.31.0

# Utilities
python-dotenv==1.0.0
